requirements.txt
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.1
django-mathfilters==1.0.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
Pillow==7.1.2
public==2019.4.13
pytz==2020.1
query-string==2019.4.13
requests==2.24.0
urllib3==1.25.9
whitenoise==5.1.0
psycopg2==2.8.5
WeasyPrint==51